refatoração de do token service + alteração de documentação
