+++ src/main/java/com/maisprati/forum/config/WebSecurityConfig.java
@@ -33,7 +33,6 @@ public class WebSecurityConfig {
                         .requestMatchers(HttpMethod.POST,"/api/auth/register").permitAll()
                         .requestMatchers(HttpMethod.POST,"/api/auth/login").permitAll()
                         .requestMatchers("/api/topics").hasRole("USER")
-                        .requestMatchers("/api/forum").hasRole("USER")
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
