+++ src/main/java/com/maisprati/forum/model/UserRole.java
@@ -1,8 +1,8 @@
 package com.maisprati.forum.model;
 
 public enum UserRole {
-    ADMIN("ADMIN"),
-    USER("USER");
+    ADMIN("admin"),
+    USER("user");
 
     private String role;
